{username/date_order} naming scheme + covers

+++ fanbox_dl/__init__.py
@@ -1,6 +1,7 @@
 import json
 import os
 import sys
+from collections import Counter
 from pathlib import Path, PurePosixPath
 from typing import Any, Optional
 from urllib.parse import unquote, urlsplit
@@ -9,15 +10,19 @@ import click
 import requests
 
 
-def download(url: str, dest_dir: Path, clobber: bool, session_id: str) -> None:
-    fullpath = unquote(urlsplit(url).path)
-    filename = PurePosixPath(fullpath).name
-
-    req = get(url, session_id)
-    file = Path(dest_dir, filename)
-    if clobber or not file.exists():
-        with open(file, "wb") as f:
-            f.write(req.content)
+def download(
+    url: str, dest: Path, clobber: bool, dry_run: bool, session_id: str
+) -> None:
+    if dry_run:
+        print(f"{url} => {dest}")
+        pass
+    else:
+        req = get(url, session_id)
+        file = Path(dest)
+        file.parent.mkdir(parents=True, exist_ok=True)
+        if clobber or not file.exists():
+            with open(file, "wb") as f:
+                f.write(req.content)
 
 
 def get(url: str, session_id: str) -> requests.Response:
@@ -68,8 +73,11 @@ def get_posts(creator: str, session_id: str) -> Any:
 @click.option("-c", "--cookie-file", required=True)
 @click.option("-o", "--output", default=".", show_default=True)
 @click.option("--clobber/--no-clobber", default=False, show_default=True)
+@click.option("--dry-run", is_flag=True, show_default=True)
 @click.argument("creator")
-def main(cookie_file: str, output: str, clobber: bool, creator: str) -> None:
+def main(
+    cookie_file: str, output: str, clobber: bool, dry_run: bool, creator: str
+) -> None:
     with open(cookie_file) as f:
         session_id = f.read().strip()
 
@@ -78,34 +86,50 @@ def main(cookie_file: str, output: str, clobber: bool, creator: str) -> None:
         print(f"Error: Couldn't fetch posts of {creator}", file=sys.stderr)
         sys.exit(1)
 
-    posts_len = len(posts)
+    get_prefix = (
+        lambda post: f'{post["creatorId"]}/{post["publishedDatetime"].split("T")[0]}'
+    )
+
+    prefix_counter = Counter()
+    for post in posts:
+        prefix_counter[get_prefix(post)] += 1
+
+    prefix_i = 0
     for i, post in enumerate(posts):
-        print(f"Fetching post {post['id']} ({i + 1}/{posts_len})", file=sys.stderr)
+        prefix = get_prefix(post)
+        if prefix_counter[prefix] > 1:
+            prefix_i += 1
+        else:
+            prefix_i = 0
+        if prefix_i:
+            prefix += f"_{prefix_i}"
+
+        print(
+            f"Fetching post {post['id']} as {prefix} ({i + 1}/{len(posts)})",
+            file=sys.stderr,
+        )
         data = get_post(post["id"], session_id)
         if data is None or "body" not in data or not data["body"]:
             print(f"Warning: Couldn't fetch post {post['id']}", file=sys.stderr)
             continue
 
-        dest_dir = Path(output, post["id"])
-        try:
-            os.mkdir(dest_dir)
-        except FileExistsError:
-            pass
+        # metadata_path = dest_dir / "metadata.json"
+        # if clobber or not metadata_path.exists():
+        #    with open(metadata_path, "w") as f:
+        #        json.dump(data, f)
 
-        metadata_path = dest_dir / "metadata.json"
-        if clobber or not metadata_path.exists():
-            with open(metadata_path, "w") as f:
-                json.dump(data, f)
+        urls = (
+            [data["coverImageUrl"]]
+            + [x["originalUrl"] for x in data["body"].get("images", [])]
+            + [x["originalUrl"] for x in data["body"].get("imageMap", {}).values()]
+            + [x["url"] for x in data["body"].get("files", [])]
+            + [x["url"] for x in data["body"].get("fileMap", {}).values()]
+        )
+        images = []
+        n_pad = len(str(len(urls)))
+        for n, url in enumerate(urls):
+            dest = Path(output, f'{prefix}_{n:0{n_pad}}.{url.split(".")[-1]}')
+            images.append({"dest": dest, "url": url})
 
-        urls = set()
-        for image in data["body"].get("images", []):
-            urls.add(image["originalUrl"])
-        for image in data["body"].get("imageMap", {}).values():
-            urls.add(image["originalUrl"])
-        for file in data["body"].get("files", []):
-            urls.add(file["url"])
-        for file in data["body"].get("fileMap", {}).values():
-            urls.add(file["url"])
-
-        for url in urls:
-            download(url, dest_dir, clobber, session_id)
+        for image in images:
+            download(image["url"], image["dest"], clobber, dry_run, session_id)
